starting realization tests for sortings
