+++ namestnikov.kirill/S6/main.cpp
@@ -5,23 +5,13 @@
 #include "shell_sort.hpp"
 #include "selection_sort.hpp"
 #include "generate_numbers.hpp"
-
-template< class Iter >
-void printData(Iter first, Iter last, std::ostream & out)
-{
-  auto temp = first;
-  ++temp;
-  for (; temp != last; ++temp)
-  {
-    out << *first << " ";
-    ++first;
-  }
-  out << *first << "\n";
-}
+#include "test_sortings.hpp"
 
 int main()
 {
   using namespace namestnikov;
   std::vector< int > vec = generateNumbers< int >(20);
+  printData(vec.begin(), vec.end(), std::cout);
+  doSelectionSort(vec.begin(), vec.size(), std::less< int >());
   printData(vec.begin(), vec.end(), std::cout);
 }
