+++ namestnikov.kirill/S6/selection_sort.hpp
@@ -2,6 +2,7 @@
 #define SELECTION_SORT_HPP
 
 #include <cstddef>
+#include <iterator>
 
 namespace namestnikov
 {
