+++ namestnikov.kirill/S6/shell_sort.hpp
@@ -2,6 +2,7 @@
 #define SHELL_SORT_HPP
 
 #include <cstddef>
+#include <iterator>
 
 namespace namestnikov
 {
