+++ namestnikov.kirill/S6/test_sortings.hpp
@@ -1,0 +1,19 @@
+#ifndef TEST_SORTINGS_HPP
+#define TEST_SORTINGS_HPP
+
+#include <ostream>
+
+template< class Iter >
+void printData(Iter first, Iter last, std::ostream & out)
+{
+  auto temp = first;
+  ++temp;
+  for (; temp != last; ++temp)
+  {
+    out << *first << " ";
+    ++first;
+  }
+  out << *first << "\n";
+}
+
+#endif
